Remove Unnecessary Controllers/Actions
